appディレクトリをsrcの中に移動
