+++ README.md
@@ -7,6 +7,12 @@ yarn install
 yarn dev
 ```
 
+## git 運用
+
+### github-flow
+
+https://atmarkit.itmedia.co.jp/ait/articles/1708/01/news015.html
+
 ## UI ライブラリ
 
 ### Material UI
@@ -52,3 +58,7 @@ APIを使用する場合、格納
 ページのファイルはこちらに格納
 
 ```
+
+## MUI のカスタム Theme を Storybook で確認できるようにする
+
+https://zenn.dev/enish/articles/ff678649ecb6d9
