+++ tsconfig.json
@@ -17,8 +17,7 @@
     "incremental": true,
     "baseUrl": "./",
     "paths": {
-      "~/app/*": ["app/*"],
-      "~/src/*": ["src/*"]
+      "~/*": ["src/*"]
     },
     "plugins": [
       {
